refactor: use BigDecimal instead of double for amount

+++ src/main/java/se/salt/foreignexchangeapi/dto/ConversionResponse.java
@@ -2,11 +2,13 @@ package se.salt.foreignexchangeapi.dto;
 
 import se.salt.foreignexchangeapi.domain.CurrencyCode;
 
+import java.math.BigDecimal;
+
 public record ConversionResponse(
         CurrencyCode baseCurrency,
         CurrencyCode wantedCurrency,
-        double amount,
-        double rate,
-        double result
+        BigDecimal amount,
+        BigDecimal rate,
+        BigDecimal result
 ) {
 }
